requirements.txt
ffmpeg-python
openai-whisper
yt-dlp
customtkinter
Pillow
youtube-transcript-api
browser-cookie3
